ws/add_consumer_groups_backend (#668)
* add rpc redis test var

* add test scenario for redis configs

* fix tests on prod

* redis continuum testing script

* Using same redis container for tests, stating in config which db index to use

* adding rpc redis test

* fix eof test

* test ony for qa

* rpc queue remopved

* test fixed

* enhanced test for redis rpc

* removing unused test

* add redis script for test

* fix

* add database to uri builder

* add env

* try test circleci env

* test circleci env

* no test

* remove BOM_TEST_REDIS_* env

* perltidy

* remove redis configure file

* remove redis_cache_server and bom_cache_server

* try to remove useless modules

* move await to Test::Helper

* install method hook

* 42

* fix test

* Revert "fix test"

This reverts commit 2d510f40de872bbf4b984434505cafb1a22c9991.

* test database index

* test redis db

* skip v6

* add comment

* tidy

* Update lib/BOM/Test.pm


* add comment

* update POD


* fix tests

* revert changes

* remove redis rpcconfig file

* keep config dir

* add Mojo::Redis2

* add test

* remove useless test

* remove env variable pod

* remove env

* move adding hook of Net::Async::Redis to INIT

* rename test

* change from get to then

* update README

* remove env

* remove env

* remove env

* remove env

* remove ws_redis stuff

* trigger circleci

* trigger circleci

* add no warnings

* no critic

* typo

* inject BOM::Test to scripts

* add env NO_PURGE_REDIS

* change system to exec to use bom::test

* trigger circleci

* trigger circleci

* trigger circleci

* trigger circleci

* trigger circleci

* trigger circleci

* typo

* load config from /etc/rmg

* trigger circleci

* trigger circleci

* remove message

* fix error: BomRpc required RedisDB too early

* change bomRpc to a script, and no purge_redis in other modules

* remove debug code

* tidy

* make script consistent with Chylli's changes

* add debug code

* add debug code

* add debug code

* start bomRpc

* typo

* debug code

* sleep to wait ready

* fix test lalalalalalalalala

* trigger circleci

* trigger circleci

* perltidy

* remove debug code

* add comment

* cache::redis->redis->flushdb

* remove debug code

* resolve conflict

* change wrong condition

* Update bin/bom_rpc_for_test.pl


* Update lib/BOM/Test.pm


* Update lib/BOM/Test.pm


* Update lib/BOM/Test.pm


* use strict warnings

* Update lib/BOM/Test/Helper.pm


* remove useless code

* add pod

* reap child gracefully

* add comment

* comment

* trigger circleci

* trigger circleci

* require bom::test

* fix exit code

* fix error

+++ lib/BOM/Test.pm
@@ -3,13 +3,12 @@ package BOM::Test;
 use strict;
 use warnings;
 use Dir::Self;
-use Cwd qw/abs_path/;
-use POSIX qw/setsid/;
+use Class::Method::Modifiers qw(install_modifier);
+use RedisDB;
+use Mojo::Redis2;
 use Path::Tiny;
 use Syntax::Keyword::Try;
-use RedisDB;
 use YAML::XS;
-use await;
 
 =head1 NAME
 
@@ -18,6 +17,7 @@ BOM::Test - Do things before test
 =head1 DESCRIPTION
 
 This module is used to prepare test environment. It should be used before any other bom modules in the test file.
+It should not use any BOM::* modules to avoid affecting other modules.
 
 =head1 Environment Variables
 
@@ -29,62 +29,9 @@ This variable will be set if test is running on qa devbox. If it is set the syst
 
 =cut
 
-=item $ENV{REDIS_CACHE_SERVER}
-
-This variable will be set if test is running on qa devbox. If it is set the Cache::RedisDB will use test redis instance instead of development.
-
-=item $ENV{BOM_TEST_REDIS_RAND}
-
-This variable will be set if test is running on qa devbox. If it is set the BOM::Config::randsrv will use test redis instance instead of development.
-
-=cut
-
-=item $ENV{BOM_TEST_REDIS_REPLICATED}
-
-This variable will be set if test is running on qa devbox. If it is set the BOM::Config::Redis and other bom services
-will use test redis instance instead of development.
-
-=item $ENV{BOM_TEST_REDIS_FEED}
-
-This variable wil be set if test is running on qa devbox. If it is set the BOM::Config::Redis::redis_feed_*() will use it.
-
-=item $ENV{BOM_TEST_REDIS_EVENTS}
-
-This variable wil be set if test is running on qa devbox. If it is set the BOM::Config::Redis::redis_events_*() will use it.
-
-=item $ENV{BOM_TEST_REDIS_QUEUE}
-
-This variable wil be set if test is running on qa devbox. If it is set the BOM::Config::Redis::redis_queue_*() will use it.
-
-=item $ENV{BOM_TEST_REDIS_TRANSACTION}
-
-This variable wil be set if test is running on qa devbox. If it is set the BOM::Config::Redis::redis_transaction_*() will use it.
-
-=item $ENV{BOM_TEST_REDIS_TRANSACTION_LIMITS}
-
-This variable wil be set if test is running on qa devbox. If it is set the BOM::Config::RedisTransactionLimits::redis_limites_*() will use it.
-
-=item $ENV{BOM_TEST_REDIS_AUTH}
-
-This variable wil be set if test is running on qa devbox. If it is set the BOM::Config::Redis::redis_auth_*() will use it.
-
-=item $ENV{BOM_TEST_REDIS_MT5_USER}
-
-This variable wil be set if test is running on qa devbox. If it is set the BOM::Config::Redis::redis_mt5_user_*() will use it.
-
-=item $ENV{BOM_TEST_REDIS_P2P}
-
-This variable wil be set if test is running on qa devbox. If it is set the BOM::Config::Redis::redis_p2p_*() will use it.
-
-=item $ENV{BOM_TEST_WS_REDIS}
-
-This variable wil be set if test is running on qa devbox. If it is set the BOM::Config::Redis::redis_ws_*() will use it.
-
 =back
 
 =cut
-
-my $config_dir;
 
 BEGIN {
     my $env = do {
@@ -114,10 +61,6 @@ BEGIN {
 
     die "wrong env. Can't run test" unless (on_qa() or on_development());
 
-    my $dir_path      = __DIR__;
-    my $test_data_dir = abs_path("$dir_path/../../data");
-    $config_dir = $test_data_dir . '/config';
-
     ## no critic (Variables::RequireLocalizedPunctuationVars)
     # set a env variable to represent we are in testing mode
     $ENV{BOM_TEST}                = 1;
@@ -126,22 +69,6 @@ BEGIN {
     $ENV{PRICING_RPC_URL}         = 'http://127.0.0.1:15006/';
 
     if (on_qa()) {
-        # Redis rand and replicated servers config
-        $ENV{BOM_TEST_REDIS_REPLICATED}         = $config_dir . '/redis-replicated.yml';
-        $ENV{BOM_TEST_REDIS_RAND}               = $config_dir . '/redis.yml';
-        $ENV{BOM_TEST_REDIS_FEED}               = $config_dir . '/redis-feed.yml';
-        $ENV{BOM_TEST_REDIS_EVENTS}             = $config_dir . '/redis-replicated.yml';
-        $ENV{BOM_TEST_REDIS_RPC}                = $config_dir . '/redis-rpc.yml';
-        $ENV{BOM_TEST_REDIS_TRANSACTION}        = $config_dir . '/redis-replicated.yml';
-        $ENV{BOM_TEST_REDIS_TRANSACTION_LIMITS} = $config_dir . '/redis-transaction-limits.yml';
-        $ENV{BOM_TEST_REDIS_AUTH}               = $config_dir . '/redis.yml';
-        $ENV{BOM_TEST_REDIS_MT5_USER}           = $config_dir . '/redis-replicated.yml';
-        $ENV{BOM_TEST_REDIS_P2P}                = $config_dir . '/redis-replicated.yml';
-        $ENV{BOM_TEST_WS_REDIS}                 = $config_dir . '/redis-replicated.yml';
-
-        # Cache redis server
-        $ENV{REDIS_CACHE_SERVER} = $ENV{BOM_CACHE_SERVER} = '127.0.1.3:6385';
-
         $ENV{DB_POSTFIX}    = '_test';
         $ENV{PGSERVICEFILE} = '/home/nobody/.pg_service_test.conf';
 
@@ -157,15 +84,57 @@ BEGIN {
     delete $ENV{PERL5OPT};
 }
 
-purge_redis();
+# Redis test database index, used by Redis 'SELECT' command to change the database in the test env.
+my $REDIS_DB_INDEX = 10;
+# To avoid running too many Redis instances on QA machines at the same time, the test script will share Redis instances with the services
+# with different database index.
+# The following piece of code will change the Redis database's index while creating instances.
+install_modifier 'RedisDB' => 'around' => 'new' => sub {
+    my ($code, $class, @args) = @_;
+    my $redis = $class->$code(@args);
+    $redis->{database} = $REDIS_DB_INDEX;
+    $redis->select($REDIS_DB_INDEX);
+    return $redis;
+};
+install_modifier 'Mojo::Redis2' => 'around' => 'new' => sub {
+    my ($code, $class, @args) = @_;
+    my $redis = $class->$code(@args);
+    $redis->url->path($REDIS_DB_INDEX);
+    return $redis;
+};
+
+# We don't want to load Future::AsyncAwait in some repo (see https://trello.com/c/gZ6Gj4Mq/8724-3-async-awaitblacklist-05 ) ,
+# So we can't load Net::Async::Redis here (which will load Future::AsyncAwait).
+# The code is added into INIT block. It will run after other modules loaded.
+# So we can test whether Net::Async::Redis is loaded. If is loaded by other module, then we apply the hook
+{
+    # There will be a warn of `Too late to run INIT block` when we start BOM::Test::Script::* in .proverc
+    # That's not a problem. Lets disable it.
+    no warnings qw(void);    ## no critic
+
+    INIT {
+        if ($INC{"Net/Async/Redis.pm"}) {
+            install_modifier 'Net::Async::Redis' => 'around' => 'connect' => sub {
+                my ($code, $redis, @args) = @_;
+                return $redis->$code(@args)->then(
+                    sub {
+                        my $res = shift;
+                        $redis->select($REDIS_DB_INDEX)->then_done($res);
+                    });
+            };
+        }
+    }
+}
+
+purge_redis() unless $ENV{NO_PURGE_REDIS};
 
 =head1 Functions
 
 =head2 purge_redis
 
-Purge redis database before running a test script. Give it a clear environment.
+Purge Redis database before running a test script. Give it a clear environment.
 
-Parameters : none
+Parameters: none
 Return: 1
 
 =cut
@@ -179,16 +148,32 @@ Return: 1
 sub purge_redis {
     my %flushed_redis;
     # Here we get configuration from yml file directly, not get redis instance from BOM::Config::Redis
-    # to avoid redis to be purged many times.
-    for my $redis_yml (path($config_dir)->children) {
-        next unless $redis_yml->basename =~ /redis/;
+    # to avoid loading BOM::Config in this file
+    my $config_dir      = '/etc/rmg';
+    my @redis_yml_files = qw(
+        redis-replicated.yml
+        redis-pricer.yml
+        redis-pricer-subscription.yml
+        redis-pricer-shared.yml
+        redis-exchangerates.yml
+        redis-feed.yml
+        redis-mt5user.yml
+        redis-events.yml
+        redis-rpc.yml
+        redis-transaction.yml
+        redis-transaction-limits.yml
+        redis-auth.yml
+        redis-p2p.yml
+        ws-redis.yml
+    );
+
+    for my $redis_yml (@redis_yml_files) {
+        $redis_yml = path("$config_dir/$redis_yml");
         my $config_content = YAML::XS::LoadFile("$redis_yml");
         for my $config_key (keys %$config_content) {
             next unless $config_key =~ /write/;
             my $config = $config_content->{$config_key};
-            # TODO will add db index in the future
             next if $flushed_redis{"$config->{host}:$config->{port}"};
-
             # Since redis services in docker depend on every repo's `services.yml`,
             # maybe some redis servers are not started.
             # Here our main goal is not test redis server but reinitialize them,
@@ -198,7 +183,8 @@ sub purge_redis {
                     host => $config->{host},
                     port => $config->{port},
                     ($config->{password} ? ('password' => $config->{password}) : ()));
-                print STDERR "$$ flushing redis $config->{host}:$config->{port}\n";
+                die "Purge redis $redis_yml failed: Database index is not $REDIS_DB_INDEX. RedisDB hook not applied ?"
+                    if $redis->{database} != $REDIS_DB_INDEX;
                 $redis->flushdb();
                 $flushed_redis{"$config->{host}:$config->{port}"} = 1;
             } catch {
